refactor(interpreter): split instruction table impls
